Updated to Rails 7.1, Ruby 3.2, and dropped support for Rails 6.1, 7.0, and Ruby 3.0
Droped Ruby 3.0 to fix rvm / Ubuntu 22 / openssl 1 incompatibility error:
> LoadError: cannot load such file -- openssl
See https://github.com/rvm/rvm/issues/5209 for more

Updated gems:

- Updated `selenium-webdriver` to latest to resolve a test error
- Updated puma gem to fix rack 3 incompatibility
- Updated mocha gem to fix Minitest incompatibility:
  > NameError: uninitialized constant MiniTest

Updated travis configuration:

- Update travis to Ubuntu jammy to fix node error:
  > node: /lib/x86_64-linux-gnu/libm.so.6: version `GLIBC_2.27' not found (required by node)
- Updated Node (in travis) to 18
- Updated Chrome installation on travis to use addon to fix version mismatch

Other fixes:

- Fixed/removed some outdated config settings
- Fixed Rubocop warning by using `delete_prefix!`
- Ignored rubocop warning for Ruby 3.1 hash syntax
- Fixed error trying to deserialize certain unsafe classes via YAML:
  > Psych::DisallowedClass: Tried to load unspecified class: ActiveSupport::TimeWithZone
- Removed `coveralls` (code coverage reporter), which is unmaintained and throwing an error:
  > OpenSSL::SSL::SSLError
  > SSL_connect returned=1 errno=0 peeraddr=104.26.4.137:443 state=error: no protocols available

Note that a few tests deserialize blank strings as an empty string in dev and
as a one character whitespace string (`"\s"`) on travis.  I tried stripping the
end-of-line spaces to make the tests behave the same in both environments, but
that didn't work for some reason.  So expect tests to pass on travis, but possibly
not in a dev env.